Light Cycles

+++ revised-code-WinAPI/Chapter 08/Light Cycles (part 2)/Light Cycles.cpp
@@ -205,7 +205,8 @@ void NewGame( )
 
    g_gameOver = FALSE;
 
-   EnableMenuItem( GetMenu( g_game->GetWindow( ) ), ( UINT ) MAKEINTRESOURCEW( IDM_GAME_NEW ), MF_GRAYED );
+   EnableMenuItem( GetMenu( g_game->GetWindow( ) ),
+                   ( UINT ) ( ULONGLONG ) MAKEINTRESOURCEW( IDM_GAME_NEW ), MF_GRAYED );
 }
 
 void UpdateCycles( )
@@ -315,7 +316,8 @@ void EndGame( LONG cycle )
 {
    g_gameOver = TRUE;
 
-   EnableMenuItem( GetMenu( g_game->GetWindow( ) ), ( UINT ) MAKEINTRESOURCEW( IDM_GAME_NEW ), MF_ENABLED );
+   EnableMenuItem( GetMenu( g_game->GetWindow( ) ),
+                   ( UINT ) ( ULONGLONG ) MAKEINTRESOURCEW( IDM_GAME_NEW ), MF_ENABLED );
 
    if ( 0 == cycle )
    {
